+++ src/bittytax/bt_types.py
@@ -6,6 +6,26 @@ from enum import Enum
 from typing import NewType
 
 
+class TaxRules(Enum):
+    UK_INDIVIDUAL = "UK Individual"
+    US_INDIVIDUAL_FIFO = "US Individual (FIFO)"
+    US_INDIVIDUAL_LIFO = "US Individual (LIFO)"
+    US_INDIVIDUAL_HIFO = "US Individual (HIFO)"
+    US_INDIVIDUAL_LOFO = "US Individual (LOFO)"
+    UK_COMPANY_JAN = "UK Company (Jan)"
+    UK_COMPANY_FEB = "UK Company (Feb)"
+    UK_COMPANY_MAR = "UK Company (Mar)"
+    UK_COMPANY_APR = "UK Company (Apr)"
+    UK_COMPANY_MAY = "UK Company (May)"
+    UK_COMPANY_JUN = "UK Company (Jun)"
+    UK_COMPANY_JUL = "UK Company (Jul)"
+    UK_COMPANY_AUG = "UK Company (Aug)"
+    UK_COMPANY_SEP = "UK Company (Sep)"
+    UK_COMPANY_OCT = "UK Company (Oct)"
+    UK_COMPANY_NOV = "UK Company (Nov)"
+    UK_COMPANY_DEC = "UK Company (Dec)"
+
+
 class TrType(Enum):
     DEPOSIT = "Deposit"
     UNSTAKE = "Unstake"
@@ -49,11 +69,40 @@ class DisposalType(Enum):
     SHORT_TERM = "Short-Term"
 
 
+class CostBasisMethod(Enum):
+    FIFO = "FIFO"
+    LIFO = "LIFO"
+    HIFO = "HIFO"
+    LOFO = "LOFO"
+
+
 class TrRecordPart(Enum):
     BUY = "Buy"
     SELL = "Sell"
     FEE = "Fee"
 
+
+TAX_RULES_US_INDIVIDUAL = [
+    TaxRules.US_INDIVIDUAL_FIFO,
+    TaxRules.US_INDIVIDUAL_LIFO,
+    TaxRules.US_INDIVIDUAL_HIFO,
+    TaxRules.US_INDIVIDUAL_LOFO,
+]
+
+TAX_RULES_UK_COMPANY = [
+    TaxRules.UK_COMPANY_JAN,
+    TaxRules.UK_COMPANY_FEB,
+    TaxRules.UK_COMPANY_MAR,
+    TaxRules.UK_COMPANY_APR,
+    TaxRules.UK_COMPANY_MAY,
+    TaxRules.UK_COMPANY_JUN,
+    TaxRules.UK_COMPANY_JUL,
+    TaxRules.UK_COMPANY_AUG,
+    TaxRules.UK_COMPANY_SEP,
+    TaxRules.UK_COMPANY_OCT,
+    TaxRules.UK_COMPANY_NOV,
+    TaxRules.UK_COMPANY_DEC,
+]
 
 BUY_TYPES = (
     TrType.DEPOSIT,